requirements.txt
#python libraries required for the project
python-dotenv
streamlit
openai # best to use this or gemini model
langchain-groq